Super_SaaS_ Burger_backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic[email]>=2.0
python-dotenv>=1.0
httpx>=0.27
passlib>=1.7
reportlab>=4.0
pillow>=10.0
python-multipart>=0.0.9
